dmaengine: stm32-dma: Replaced cancel_work_sync instead of flush_work
In stm32_dma_mdma_drain it was necessary to replace flush_work with
cancel_work_sync because this api waits till the execution of the work
queue is finish. We should wait for the execution of the work to finish
when it is cancelled as the worker thread can execute in another context.

Change-Id: I80690a01fc15d64f7c40eb724a638b8b79970f09
Signed-off-by: Sandhya Sharma <sandhya.sharma@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/405535
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Amelie DELAUNAY <amelie.delaunay@foss.st.com>